Fix version in release 0.45.0
The upstream repo released v0.45.0 at the same commit as the (unreleased) v0.44.4. Having the two tags throws off the `build.sh` script, so that the resulting internal version value would be wrong (although the same code is built).

This workaround overrides the version parsed from the git tag during the build script with the version provided by CI.